requirements.txt
requests
pandas
gspread
google-auth
python-dotenv
schedule
facebook-business
shopify-python-api
openpyxl
google-auth-oauthlib